api/requirements.txt
eospy
Flask
peewee
gunicorn
supervisor
psycopg2-binary
google-cloud-translate
